add convolution_float

+++ simd_benchmarking/convolution_simd_for_dummies.c
@@ -2,6 +2,9 @@
 #include "gradient_magnitude.h"
 #include <immintrin.h>
 
+// Vers'ao intermediaria para o sind que ajuda a visualizar o vertor de soma e pixel
+// a ideia aqui eh carregar 16 pixels de uma vez
+
 Matrix conv2d(Matrix img, Matrix kernel) {
     Matrix out = {
         img.width,
@@ -12,16 +15,16 @@ Matrix conv2d(Matrix img, Matrix kernel) {
     for (int i = 0; i < out.height; i++) {
         for (int j = 0; j < out.width; j++) {
             //__m256i sum = _mm256_setzero_si256(); // sum = 0
-            int16_t sum[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0};
+            int32_t sum[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0};
             for (int k = 0; k < kernel.height; k++) {
                 for (int l = 0; l < kernel.width; l++) {
                     int x = j + l;
                     int y = i + k;
 
-                    int16_t weight = kernel.data[k*kernel.width + l];
-                    int16_t pixel = img.data[y*img.width + x];
+                    int32_t weight = kernel.data[k*kernel.width + l];
+                    int32_t pixel = img.data[y*img.width + x];
 
-                    int16_t pixeis[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0};
+                    int32_t pixeis[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0};
                     for (int m = 0; m < 16; m++) {
                         pixeis[m] = img.data[y*img.width + x + m];
                     }
